ctrlp statusline - reviewed
